# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me/avatar": {
      "put": {
        "operationId": "MediaController_uploadAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Avatar replaced.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaMutationEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "summary": "Upload and replace the authenticated user avatar",
        "tags": ["media"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "One JPEG, PNG, or WebP image, at most 10,000,000 bytes."
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "MediaController_deleteAvatar",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Avatar deleted."
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "summary": "Delete the authenticated user avatar",
        "tags": ["media"]
      }
    },
    "/api/v1/participants/{participantId}/avatar": {
      "get": {
        "operationId": "MediaController_participantAvatar",
        "parameters": [
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          },
          {
            "name": "v",
            "in": "query",
            "required": true,
            "description": "Immutable media identifier from the URL returned by the upload or resource API.",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized versioned private participant avatar.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": ["private, no-store"]
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "enum": ["Cookie"]
                }
              },
              "X-Content-Type-Options": {
                "schema": {
                  "type": "string",
                  "enum": ["nosniff"]
                }
              }
            },
            "content": {
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Read an authorized versioned participant avatar",
        "tags": ["media"]
      }
    },
    "/api/v1/groups/{groupId}/image": {
      "put": {
        "operationId": "MediaController_uploadGroupImage",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group image replaced.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaMutationEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "summary": "Upload and replace a group image as owner or administrator",
        "tags": ["media"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "One JPEG, PNG, or WebP image, at most 10,000,000 bytes."
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "MediaController_deleteGroupImage",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Group image deleted."
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "summary": "Delete a group image as owner or administrator",
        "tags": ["media"]
      },
      "get": {
        "operationId": "MediaController_groupImage",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          },
          {
            "name": "v",
            "in": "query",
            "required": true,
            "description": "Immutable media identifier from the URL returned by the upload or resource API.",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized versioned private group image.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": ["private, no-store"]
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "enum": ["Cookie"]
                }
              },
              "X-Content-Type-Options": {
                "schema": {
                  "type": "string",
                  "enum": ["nosniff"]
                }
              }
            },
            "content": {
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Read an authorized versioned group image",
        "tags": ["media"]
      }
    }
  },
  "info": {
    "title": "SPLITO API",
    "description": "Versioned API for SPLITO expense sharing and its financial journal",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current SPLITO deployment"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SPLITO_SESSION"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token"
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
                "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
              },
              "fieldErrors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["field", "message"],
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "MediaMutation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["url", "version"],
        "properties": {
          "url": {
            "type": "string",
            "description": "Cookie-authenticated, versioned, same-origin image URL."
          },
          "version": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          }
        }
      },
...
      "MediaMutationEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MediaMutation"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed input or identifier.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "A valid session is required.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated identity is not authorized.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist or is not visible to this identity.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "The resource version, idempotency body, or financial state conflicts.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PreconditionFailed": {
        "description": "The supplied resource version is stale.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The uploaded file or multipart request exceeds its limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "The upload is not an accepted image or multipart body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "The financial-domain invariant was not satisfied.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "A required optimistic-concurrency header is missing.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "RateLimited": {
        "description": "The request rate limit was exceeded.",
        "headers": {
          "Retry-After": {
            "description": "Minimum seconds before retrying; 60 for cooldown or up to 900 for quota.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 900
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "BadGateway": {
        "description": "The external SMS provider did not accept the invitation message.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A required external delivery adapter is unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}